pyproject.toml
[tool.poetry]
name = "gifbot"
version = "0.1.0"
description = "Create animations and memes from videos/gifs and text"
authors = ["Matheus Fillipe <matheusfillipeag@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-telegram-bot = "^20.6"
python-dotenv = "^1.0.0"


[tool.poetry.group.dev.dependencies]
pre-commit = "^3.5.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
target-version = ['py311']

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 120
force_single_line = true
profile = "black"

[tool.docformatter]
wrap-summaries = 120
wrap-descriptions = 120
in-place = true

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.ruff]
line-length = 120
